Añadiendo cambios

+++ Tema_5/Ejmplo/Amazon/Amazon_Montecastelo/Controllers/AmazonController.cs
@@ -4,6 +4,8 @@ using Amazon_Montecastelo.Database;
 using LoboGarcesWindarTarea3;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using Microsoft.AspNetCore.Authorization;
+using Amazon_Montecastelo.Models;
 
 namespace Amazon_Montecastelo.Controllers
 {
@@ -12,14 +14,14 @@ namespace Amazon_Montecastelo.Controllers
         private readonly ProductoRepository _productoRepositorio;
         private readonly CarritoRepository _carritoRepositorio;
         private readonly VentaRepository _ventaRepository;
-  
-        public AmazonController(Conexion context )
+
+        public AmazonController(Conexion context)
         {
 
             _productoRepositorio = new ProductoRepository(context);
-            _carritoRepositorio= new CarritoRepository(context);
-            _ventaRepository= new VentaRepository(context);
-        
+            _carritoRepositorio = new CarritoRepository(context);
+            _ventaRepository = new VentaRepository(context);
+
 
         }
         [Route("/")]
@@ -32,10 +34,17 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "administrador")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
+
             var productos = await _productoRepositorio.GetAllProducto();
 
-
             return View(productos);
+
         }
         [Route("/Amazon/agregar")]
         public IActionResult Agregar()
@@ -43,8 +52,14 @@ namespace Amazon_Montecastelo.Controllers
             if (!GlobalInfo.IsLogged)
             {
                 return View(GlobalInfo.LoginView);
-            }
 
+            }
+            if (GlobalInfo.UsuarioLogeado.userType != "administrador")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
             return View(new Productos());
         }
 
@@ -57,15 +72,22 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+
             if (producto.ProductoID > 0)
             {
-                
+
                 await _productoRepositorio.UpdateProducto(producto);
             }
             else
             {
-               
+
                 await _productoRepositorio.CreateProducto(producto);
+            }
+            if (GlobalInfo.UsuarioLogeado.userType != "administrador")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
             }
 
             return RedirectToAction("Producto");
@@ -80,7 +102,12 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
-
+            if (GlobalInfo.UsuarioLogeado.userType != "administrador")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
             var producto = await _productoRepositorio.GetProducto(id);
 
             return View("Agregar", producto);
@@ -95,20 +122,40 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "administrador")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
             await _productoRepositorio.DeleteProducto(id);
 
             return RedirectToAction("Producto");
         }
 
         [Route("/Amazon/ProductoUsuario")]
+
         public async Task<IActionResult> ProductoUsuario()
         {
             if (!GlobalInfo.IsLogged)
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "usuario")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
+
             var productos = await _productoRepositorio.GetAllProducto();
-            return View(productos);
+            var viewModel = new ProductoViewModel
+            {
+                Productos = productos
+
+            };
+
+            return View(viewModel);
         }
 
         [HttpPost]
@@ -119,9 +166,15 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "usuario")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
             var userId = GlobalInfo.UsuarioLogeado.UsuarioID;
 
-            var producto =  await _productoRepositorio.GetProducto(productoId);
+            var producto = await _productoRepositorio.GetProducto(productoId);
 
             var carrito = GlobalInfo.MiCarrito;
 
@@ -137,7 +190,8 @@ namespace Amazon_Montecastelo.Controllers
                     {
                         detalle.Cantidad++;
                         detalle.PrecioTotal = detalle.Cantidad * producto.Precio;
-                        
+
+
                     }
                     else if (accion == "decrementar" && detalle.Cantidad > 1)
                     {
@@ -151,18 +205,20 @@ namespace Amazon_Montecastelo.Controllers
             {
                 carrito.Detalles.Add(new DetalleCarrito
                 {
+
                     ProductoID = producto.ProductoID,
                     Cantidad = 1,
                     PrecioTotal = producto.Precio,
                     PrecioUnitario = producto.Precio,
-                  
-            });
+
+                });
             }
 
             carrito.TotalVenta = carrito.Detalles.Sum(x => x.PrecioTotal);
-           
+
+
             await _carritoRepositorio.Guardar(carrito);
-         
+
             return RedirectToAction("Carrito", "Amazon");
         }
 
@@ -173,13 +229,20 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "usuario")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
             var userId = GlobalInfo.UsuarioLogeado.UsuarioID;
 
             var carrito = await _carritoRepositorio.ObtenerCarrito(userId);
-           
+
             carrito.Subtotal = carrito.Detalles.Sum(x => x.PrecioTotal);
 
             carrito.CantidaProducto = carrito.Detalles.Sum(x => x.Cantidad);
+
 
             return View(carrito);
         }
@@ -192,13 +255,19 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
-          await _carritoRepositorio.EliminarProdcutoDelCarrito(productoId);
+            if (GlobalInfo.UsuarioLogeado.userType != "usuario")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
+            await _carritoRepositorio.EliminarProdcutoDelCarrito(productoId);
 
             return RedirectToAction("Carrito", "Amazon");
         }
 
 
-    
+
         [Route("/Amazon/Venta")]
         public async Task<IActionResult> Venta()
         {
@@ -206,6 +275,14 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "usuario")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+
+            }
+
             var userId = GlobalInfo.UsuarioLogeado.UsuarioID;
 
             var venta = await _ventaRepository.ObtenerVenta(userId);
@@ -221,27 +298,34 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "usuario")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
             var userId = GlobalInfo.UsuarioLogeado.UsuarioID;
 
-           
+
             var producto = await _productoRepositorio.GetProducto(productoId);
 
-           
+
             var carrito = await _carritoRepositorio.ObtenerCarrito(userId);
 
             var venta = await _ventaRepository.ObtenerVenta(userId);
 
-         
+
             if (venta == null)
             {
                 venta = new Venta
                 {
-                    
+                    FechaVenta = DateTime.Now,
                     Detalles = new List<DetalleVenta>()
                 };
             }
             if (accion == "tramitarPedido")
             {
+
 
                 foreach (var detalleCarrito in carrito.Detalles)
                 {
@@ -250,7 +334,9 @@ namespace Amazon_Montecastelo.Controllers
                         ProductoID = detalleCarrito.ProductoID,
                         Cantidad = detalleCarrito.Cantidad,
                         PrecioTotal = detalleCarrito.PrecioTotal,
-                        PrecioUnitario = detalleCarrito.PrecioUnitario,
+                        PrecioUnitario = detalleCarrito.PrecioUnitario
+
+
 
                     };
 
@@ -258,11 +344,12 @@ namespace Amazon_Montecastelo.Controllers
                 }
             }
 
-          
+
             venta.TotalVenta = venta.Detalles.Sum(x => x.PrecioTotal);
 
-          
+
             await _ventaRepository.Guardar(venta);
+
             await _carritoRepositorio.EliminarDelCarrito(carrito.CarritoID);
 
             return RedirectToAction("Venta", "Amazon");
@@ -277,9 +364,41 @@ namespace Amazon_Montecastelo.Controllers
             {
                 return View(GlobalInfo.LoginView);
             }
+            if (GlobalInfo.UsuarioLogeado.userType != "usuario")
+            {
+                // Redirige o muestra un mensaje de error porque el usuario no es un administrador
+                TempData["MensajeAccesoDenegado"] = "No tienes permisos para acceder a esta página.";
+                return RedirectToAction("AccesoDenegado");
+            }
             await _ventaRepository.EliminarDelaVenta(productoId);
 
             return RedirectToAction("Venta", "Amazon");
+        }
+
+        [HttpPost]
+        [Route("/Amazon/Filtrar")]
+        public async Task<IActionResult> Filtrar(string nombre,decimal precioDesde, decimal precioHasta)
+        {
+            var filtrado = await _productoRepositorio.Filtrar(nombre,precioDesde, precioHasta);
+            var viewModel = new ProductoViewModel
+            {
+                Productos = filtrado,
+              
+
+            };
+
+            return PartialView("ProductoUsuario",  viewModel);
+        }
+
+
+        [Route("/Amazon/AccesoDenegado")]
+
+        public IActionResult AccesoDenegado()
+        {
+
+            var mensaje = TempData["MensajeAccesoDenegado"] as string;
+            ViewBag.MensajeAccesoDenegado = mensaje;
+            return View();
         }
     }
